Backend/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
apturl==0.5.2
attrs==23.2.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
cowsay==6.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.1.2
fabric==3.2.2
fasteners==0.14.1
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.18.2
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==3.6
importlib-metadata==4.6.4
importlib_resources==6.4.0
invoke==2.2.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.2.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2024.0.1
platformdirs==4.2.2
protobuf==4.21.12
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.19.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
systemd-python==234
typing_extensions==4.12.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.26.3
wadllib==1.3.6
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
